Issue #12 - Crash on clear cache

+++ UniversalImageLoader/src/com/nostra13/universalimageloader/cache/disc/DiscCache.java
@@ -25,8 +25,10 @@ public abstract class DiscCache {
 	/** Clears cache directory */
 	public void clear() {
 		File[] files = cacheDir.listFiles();
-		for (File f : files) {
-			f.delete();
+		if (files != null) {
+			for (File f : files) {
+				f.delete();
+			}
 		}
 	}
